requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
dnspython>=2.4.2
rich>=13.7.0
colorama>=0.4.6